text_scoring.py added
